395-time-series-rdb-connection: moved tests from TimeSeriesRDBClientWithoutConnTest.java to TimeSeriesRDBClientTest
