+++ JPS_BASE_LIB/src/test/java/uk/ac/cam/cares/jps/base/timeseries/TimeSeriesRDBClientTest.java
@@ -1,18 +1,28 @@
 package uk.ac.cam.cares.jps.base.timeseries;
 
-import org.jooq.SQLDialect;
 import org.junit.Assert;
+import org.junit.ClassRule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 
+import uk.ac.cam.cares.jps.base.exception.JPSRuntimeException;
+
+import java.io.FileWriter;
+import java.io.IOException;
 import java.lang.reflect.Field;
+import java.nio.file.Paths;
 import java.time.Instant;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * This class provides unit tests for the TimeSeriesRDBClient class
  */
 
 public class TimeSeriesRDBClientTest {
-
+    @ClassRule
+    public static TemporaryFolder folder = new TemporaryFolder();
+    
     @Test
     public void testConstructor() throws NoSuchFieldException, SecurityException, IllegalArgumentException, IllegalAccessException {
         TimeSeriesRDBClient<Instant> client = new TimeSeriesRDBClient<>(Instant.class);
@@ -26,41 +36,103 @@ public class TimeSeriesRDBClientTest {
     }
 
     @Test
-    public void testPrivateDatabaseRelatedFields() throws NoSuchFieldException, SecurityException, IllegalArgumentException, IllegalAccessException {
+    public void testSetAndGetRdbURL() {
         TimeSeriesRDBClient<Instant> client = new TimeSeriesRDBClient<>(Instant.class);
-        // Retrieve the value of the private field 'dialect' of the client to check its value
-        Field dialectField = client.getClass().getDeclaredField("DIALECT");
-        dialectField.setAccessible(true);
-        SQLDialect dialect = (SQLDialect) dialectField.get(client);
-        Assert.assertEquals(SQLDialect.POSTGRES, dialect);
-        // Retrieve the value of the private field 'dbTableName' of the client to check its value
-        Field tableNameField = client.getClass().getDeclaredField("DB_TABLE_NAME");
-        tableNameField.setAccessible(true);
-        String tableName = (String) tableNameField.get(client);
-        Assert.assertEquals("dbTable", tableName);
-        // Retrieve the value of the private field 'dataIRIcolumn' of the client to check its value
-        Field dataIRIcolumnField = client.getClass().getDeclaredField("DATA_IRI_COLUMN");
-        dataIRIcolumnField.setAccessible(true);
-        org.jooq.Field<String> dataIRIcolumn = (org.jooq.Field<String>) dataIRIcolumnField.get(client);
-        Assert.assertEquals("dataIRI", dataIRIcolumn.getName());
-        Assert.assertEquals(String.class, dataIRIcolumn.getType());
-        // Retrieve the value of the private field 'tsIRIcolumn' of the client to check its value
-        Field tsIRIcolumnField = client.getClass().getDeclaredField("TS_IRI_COLUMN");
-        tsIRIcolumnField.setAccessible(true);
-        org.jooq.Field<String> tsIRIcolumn = (org.jooq.Field<String>) tsIRIcolumnField.get(client);
-        Assert.assertEquals("timeseriesIRI", tsIRIcolumn.getName());
-        Assert.assertEquals(String.class, tsIRIcolumn.getType());
-        // Retrieve the value of the private field 'tsTableNameColumn' of the client to check its value
-        Field tsTableNameColumnField = client.getClass().getDeclaredField("TABLENAME_COLUMN");
-        tsTableNameColumnField.setAccessible(true);
-        org.jooq.Field<String> tsTableNameColumn = (org.jooq.Field<String>) tsTableNameColumnField.get(client);
-        Assert.assertEquals("tableName", tsTableNameColumn.getName());
-        Assert.assertEquals(String.class, tsTableNameColumn.getType());
-        // Retrieve the value of the private field 'columnNameColumn' of the client to check its value
-        Field columnNameColumnField = client.getClass().getDeclaredField("COLUMNNAME_COLUMN");
-        columnNameColumnField.setAccessible(true);
-        org.jooq.Field<String> columnNameColumn = (org.jooq.Field<String>) columnNameColumnField.get(client);
-        Assert.assertEquals("columnName", columnNameColumn.getName());
-        Assert.assertEquals(String.class, columnNameColumn.getType());
+        Assert.assertNull(client.getRdbURL());
+        client.setRdbURL("http://localhost:5342");
+        Assert.assertEquals("http://localhost:5342", client.getRdbURL());
+    }
+
+    @Test
+    public void testSetAndGetRdbUser() {
+        TimeSeriesRDBClient<Instant> client = new TimeSeriesRDBClient<>(Instant.class);
+        Assert.assertNull(client.getRdbUser());
+        client.setRdbUser("postgres");
+        Assert.assertEquals("postgres", client.getRdbUser());
+    }
+
+    @Test
+    public void testSetRdbPassword() throws NoSuchFieldException, SecurityException, IllegalArgumentException, IllegalAccessException {
+        TimeSeriesRDBClient<Instant> client = new TimeSeriesRDBClient<>(Instant.class);
+        // Retrieve the value of the private field 'rdbPassword' of the client to check its value
+        Field rdbPasswordField = client.getClass().getDeclaredField("rdbPassword");
+        rdbPasswordField.setAccessible(true);
+
+        Assert.assertNull(rdbPasswordField.get(client));
+        client.setRdbPassword("password");
+        Assert.assertNotNull(rdbPasswordField.get(client));
+        Assert.assertEquals("password", rdbPasswordField.get(client));
+    }
+
+    @Test
+    public void testLoadRdbConfig () throws IOException {
+        TimeSeriesRDBClient<Instant> client = new TimeSeriesRDBClient<>(Instant.class);
+        // Filepath to not yet created file in temporary test folder
+        String filepath = Paths.get(folder.getRoot().toString(), "timeseries.properties").toString();
+        // JPSRuntime error messages
+        String m1 = "TimeSeriesRDBClient: No properties file found at specified filepath: " + filepath;
+        String m2 = "TimeSeriesRDBClient: Properties file is missing \"db.url=<rdb_url>\" ";
+        String m3 = "TimeSeriesRDBClient: Properties file is missing \"db.user=<rdb_username>\" ";
+        String m4 = "TimeSeriesRDBClient: Properties file is missing \"db.password=<rdb_password>\" ";
+
+        // Test for non-existing properties file
+        try {
+            client.loadRdbConfigs(filepath);
+            Assert.fail();
+        } catch (JPSRuntimeException e) {
+            Assert.assertEquals(m1, e.getMessage());
+        }
+
+        // Test for missing Rdb URL by creating a file only containing user and password
+        writePropertyFile(filepath, Arrays.asList("db.user=test_user", "db.password=test_password"));
+        // Try loading RDB configs
+        try {
+            client.loadRdbConfigs(filepath);
+            Assert.fail();
+        } catch (JPSRuntimeException e) {
+            Assert.assertEquals(m2, e.getMessage());
+        }
+
+        // Test for missing user name by creating a file only containing url and password
+        writePropertyFile(filepath, Arrays.asList("db.url=test_url", "db.password=test_password"));
+        // Try loading RDB configs
+        try {
+            client.loadRdbConfigs(filepath);
+            Assert.fail();
+        } catch (JPSRuntimeException e) {
+            Assert.assertEquals(m3, e.getMessage());
+        }
+
+        // Test for missing password  by creating a file only containing url and user
+        writePropertyFile(filepath, Arrays.asList("db.url=test_url", "db.user=test_user"));
+        // Try loading RDB configs
+        try {
+            client.loadRdbConfigs(filepath);
+            Assert.fail();
+        } catch (JPSRuntimeException e) {
+            Assert.assertEquals(m4, e.getMessage());
+        }
+
+        // Test for proper URL, username and password
+        writePropertyFile(filepath, Arrays.asList("db.url=test_url", "db.user=test_user", "db.password=test_password"));
+        // Try loading RDB configs
+        try {
+            client.loadRdbConfigs(filepath);
+        } catch (Exception e) {
+            Assert.fail(e.getMessage());
+        }
+        Assert.assertEquals("test_url", client.getRdbURL());
+        Assert.assertEquals("test_user", client.getRdbUser());
+    }
+
+    private void writePropertyFile(String filepath, List<String> properties) throws IOException {
+        // Overwrite potentially existing properties file
+        FileWriter writer = new FileWriter(filepath, false);
+        // Populate file
+        for (String s : properties) {
+            writer.write(s + "\n");
+        }
+        // Close the file and return the file
+        writer.close();
     }
 }
